+++ src/interface/enum.ts
@@ -1,0 +1,7 @@
+/** 请求方法 */
+export enum ReqMethodName {
+    GET = 'get',
+    DELETE = 'delete',
+    PUT = 'put',
+    POST = 'post'
+}
